requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
phonenumbers==8.13.25
pillow==11.2.1
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
setuptools==80.8.0
sqlparse==0.5.3
